Make sure the toolchain is mounted as exec.
If someone bind mounts a toolchain into /build/cos-tools from a COS
host, chances are that the bind mount will be noexec (since the source
mount of that bind mount will likely be noexec). Since we need to
execute the toolchain, let's remount it as exec.

Tested by running cos-gpu-installer with a preloaded toolchain in
/var/lib/cos-tools, bind mounted inside the container to
/build/cos-tools.