Break dependency on host's "get_metadata_value" script
cos-gpu-installer currently depends on the "get_metadata_value" script
from the COS host environment. This causes problems when
cos-gpu-installer needs to run on COS images that do not have this
installed. Since this is a simple enough and stable enough script, let's
just package it into the cos-gpu-installer Docker image to avoid taking
this dependency on the host.

The script is copied from
https://cos.googlesource.com/cos/overlays/board-overlays/+/aaf6944669766f674f0523e5f875d11c9f0c78e1/project-lakitu/app-admin/google-guest-agent/files/get_metadata_value